% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{convert_fragments_gr}
\alias{convert_fragments_gr}
\title{Convert a list of fragment data.table to GenomicRanges}
\usage{
convert_fragments_gr(fragments, n_threads = 1)
}
\arguments{
\item{fragments}{A}

\item{n_threads}{}
}
\description{
Convert a list of fragment data.table to GenomicRanges
}
