% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim_reduction.R
\name{jaccard}
\alias{jaccard}
\title{Compute Jaccard distances}
\usage{
jaccard(m)
}
\arguments{
\item{m}{a sparse matrix of KNN}
}
\value{
a sparse matrix of Jaccard distances.
}
\description{
Compute Jaccard distances
}
