% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library_diversity.R
\name{tenx_atac_freq}
\alias{tenx_atac_freq}
\title{Compute a frequency and metrics table for a 10x ATAC-seq experiment}
\usage{
tenx_atac_freq(fragments_tsv, keep_barcodes = NULL, summary_csv)
}
\arguments{
\item{fragments_tsv}{path to an outs/fragments.tsv.gz file generated by cellranger-atac}

\item{keep_barcodes}{a character vector specifying which barcodes to retain for analysis}

\item{summary_csv}{path to an outs/summary.csv file generated by cellranger-atac}
}
\value{
a list with two objects: freq_mat, a 2-column matrix containing the histogram of fragment counts; metrics, a data.frame with totals for reads, umis, and counts.
}
\description{
Compute a frequency and metrics table for a 10x ATAC-seq experiment
}
